+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Features
 
 - Adds an event system for ECS operations, like entity creation/removal and component addition/removal (#330, #331, #333, #342, #344)
-- Adds custom event support for the event system (#340)
+- Adds custom event support for the event system (#340, #348)
 - Adds method `World.Shrink` for freeing memory that exceeds current requirements (#323)
 
 ### Performance